refactor: :recycle: move example package to examples.py
